requirements2.txt


linear-tsv==1.1.0
llvmlite @ file:///C:/b/abs_a8i9keuf6p/croot/llvmlite_1683555140340/work
lmfit==1.2.1

matplotlib==3.7.2
matplotlib-inline @ file:///C:/ci/matplotlib-inline_1661915841596/work


numexpr==2.8.4
numpy==1.24.1
numpydoc @ file:///C:/b/abs_cfdd4zxbga/croot/numpydoc_1668085912100/work

opencv-contrib-python==4.6.0.66
opencv-python==4.7.0.72
packaging @ file:///C:/b/abs_ed_kb9w6g4/croot/packaging_1678965418855/work
pandas==1.5.3
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
Pillow==9.4.0

powerxrd==2.2.0

scikit-image @ file:///C:/ci/scikit-image_1648214340990/work
scikit-learn @ file:///C:/ci/scikit-learn_1658344472767/work
scikit-learn-intelex==2021.20220215.102710
scipy==1.10.1
streamlit==1.24.1
sympy @ file:///C:/b/abs_95fbf1z7n6/croot/sympy_1668202411612/work
